update file name - now we add the _logger end in each file
